+++ tests/integration_tests/test1.py
@@ -7,7 +7,7 @@ from easycrystallography.Components.Susceptibility import MagneticSusceptibility
 
 from easydiffraction import Phase
 from easydiffraction.calculators.wrapper_factory import WrapperFactory as Calculator
-from easydiffraction.Profiles.P1D import Instrument1DCWParameters
+from easydiffraction.job.experiment.pd_1d import Instrument1DCWParameters
 from easydiffraction.sample import Sample
 
 calculator = Calculator()
