+++ tests/integration_tests/test3.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from easydiffraction import Phases
 from easydiffraction.calculators.wrapper_factory import WrapperFactory
-from easydiffraction.Profiles.P1D import Instrument1DCWParameters
+from easydiffraction.job.experiment.pd_1d import Instrument1DCWParameters
 from easydiffraction.sample import Sample
 
 i = WrapperFactory()
